implemented buttons for italic,bold, underline plugins
