+++ src/components/RichTextEditor/EditorPlugins/Plugins/StrikethroughPlugin/index.ts
@@ -1,3 +1,16 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-export * from './StrikethroughPlugin';
+import { MARK_STRIKETHROUGH, createStrikethroughPlugin } from '@udecode/plate';
+import { StrikethroughElement } from './StrikethroughElement';
+import { Plugin } from '../types';
+import { StrikethroughButton } from './StrikethroughButton';
+
+export class StrikethroughPlugin implements Plugin {
+    public id = MARK_STRIKETHROUGH;
+    public element = StrikethroughElement;
+    public button = StrikethroughButton;
+
+    plugins() {
+        return [createStrikethroughPlugin()];
+    }
+}
